# 54remitflow-unified-platform/docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Remittance Platform API
  description: |
    Comprehensive API for the Remittance Platform supporting:
    - Agent management and hierarchy
    - Customer accounts and KYC
    - Transaction processing (QR payments, P2P, cash in/out)
    - Multi-bank routing
    - Float management
    - Commission settlement
    - Mojaloop integration
  version: 2.0.0
  contact:
    name: API Support
    email: api-support@remittance-platform.com
  license:
    name: Proprietary
    url: https://remittance-platform.com/license
servers:
  - url: https://api.remittance-platform.com/v1
    description: Production
  - url: https://staging-api.remittance-platform.com/v1
    description: Staging
  - url: http://localhost:8000/v1
    description: Development
security:
  - bearerAuth: []
  - apiKeyAuth: []
...
paths:
  /auth/login:
    post:
      tags: [Authentication]
      summary: Authenticate user
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  example: agent@example.com
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Successful authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
  schemas:
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          default: Bearer
        expires_in:
          type: integer
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...