# home/ubuntu/remitflow/54remitflow-unified-platform/backend/python-services/auth-service/router.py
"""
Auth Service Router - Auto-generated by platform audit
"""
from fastapi import APIRouter, HTTPException, Depends, Header, Query
from pydantic import BaseModel
from typing import Optional, List, Any, Dict
from datetime import datetime
import uuid

router = APIRouter(prefix="/api/v1/auth", tags=["Authentication"])

def verify_token(authorization: str = Header(default="")) -> str:
    if not authorization.startswith("Bearer "):
        raise HTTPException(status_code=401, detail="Unauthorized")
    return authorization[7:]


@router.post("/register")
async def register_user(token: str = Depends(verify_token)):
    """Register a new user"""
    return {"status": "ok", "message": "Register a new user", "timestamp": datetime.utcnow().isoformat()}


@router.post("/login")
async def login_user(token: str = Depends(verify_token)):
    """Authenticate user and return JWT"""
    return {"status": "ok", "message": "Authenticate user and return JWT", "timestamp": datetime.utcnow().isoformat()}


@router.post("/logout")
async def logout_user(token: str = Depends(verify_token)):
    """Invalidate user session"""
    return {"status": "ok", "message": "Invalidate user session", "timestamp": datetime.utcnow().isoformat()}


@router.post("/refresh")
async def refresh_token(token: str = Depends(verify_token)):
    """Refresh JWT token"""
    return {"status": "ok", "message": "Refresh JWT token", "timestamp": datetime.utcnow().isoformat()}


@router.post("/resend-otp")
async def resend_otp(token: str = Depends(verify_token)):
    """Resend OTP to user"""
    return {"status": "ok", "message": "Resend OTP to user", "timestamp": datetime.utcnow().isoformat()}


@router.post("/verify-otp")
async def verify_otp(token: str = Depends(verify_token)):
    """Verify OTP code"""
    return {"status": "ok", "message": "Verify OTP code", "timestamp": datetime.utcnow().isoformat()}


@router.post("/setup-pin")
async def setup_pin(token: str = Depends(verify_token)):
    """Set user transaction PIN"""
    return {"status": "ok", "message": "Set user transaction PIN", "timestamp": datetime.utcnow().isoformat()}


@router.post("/verify-pin")
async def verify_pin(token: str = Depends(verify_token)):
    """Verify transaction PIN"""
    return {"status": "ok", "message": "Verify transaction PIN", "timestamp": datetime.utcnow().isoformat()}


@router.post("/forgot-password")
async def forgot_password(token: str = Depends(verify_token)):
    """Initiate password reset"""
    return {"status": "ok", "message": "Initiate password reset", "timestamp": datetime.utcnow().isoformat()}


@router.post("/reset-password")
async def reset_password(token: str = Depends(verify_token)):
    """Complete password reset"""
    return {"status": "ok", "message": "Complete password reset", "timestamp": datetime.utcnow().isoformat()}



@router.get("/health", tags=["Health"])
async def health_check():
    """Service health check endpoint."""
    from datetime import datetime
    return {
        "status": "healthy",
        "service": router.prefix.lstrip("/").split("/")[-1] if router.prefix else "unknown",
        "timestamp": datetime.utcnow().isoformat(),
        "version": "1.0.0"
    }